+++ frontend/svelte/src/lib/components/launchpad/ProposalCard.svelte
