+++ frontend/svelte/src/lib/components/launchpad/Proposals.svelte
