+++ frontend/svelte/src/lib/components/launchpad/VotesProgress.svelte
