+++ frontend/svelte/src/lib/components/project-detail/CommitmentProgressBar.svelte
