+++ frontend/svelte/src/lib/components/project-detail/ProjectInfoSection.svelte
