+++ frontend/svelte/src/lib/stores/projects.store.ts
