+++ frontend/svelte/src/routes/ProjectDetail.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { onDestroy, onMount } from "svelte";
-  import ProjectInfoSection from "../lib/components/sns-project-detail/ProjectInfoSection.svelte";
-  import ProjectStatusSection from "../lib/components/sns-project-detail/ProjectStatusSection.svelte";
+  import ProjectInfoSection from "../lib/components/project-detail/ProjectInfoSection.svelte";
+  import ProjectStatusSection from "../lib/components/project-detail/ProjectStatusSection.svelte";
   import TwoColumns from "../lib/components/ui/TwoColumns.svelte";
   import { IS_TESTNET } from "../lib/constants/environment.constants";
   import { AppPath } from "../lib/constants/routes.constants";
@@ -19,7 +19,7 @@
   import {
     type SnsFullProject,
     snsFullProjectStore,
-  } from "../lib/stores/snsProjects.store";
+  } from "../lib/stores/projects.store";
   import { getSnsProjectById } from "../lib/utils/sns.utils";
   import Spinner from "../lib/components/ui/Spinner.svelte";
